P01-the-loop/backend/pyproject.toml
[project]
name = "mini-deerflow-p01"
version = "0.1.0"
description = "mini-deerflow stage P01: The Loop"
requires-python = ">=3.12"
dependencies = [
    "debugpy>=1.8.0",
    "langchain>=1.2.3",
    "langgraph>=1.0.6,<1.0.10",
    "langgraph-cli[inmem]>=0.4.0,<0.5.0",
    "langchain-openai>=1.1.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["packages/harness/mini_deerflow"]
